--- skore/tests/unit/sklearn/train_test_split/warning/test_high_class_imbalance_warning.py
@@ -1,28 +1,0 @@
-import numpy
-import pandas
-import pytest
-
-from skore._sklearn.train_test_split.warning import (
-    HighClassImbalanceWarning,
-)
-
-target = [0] * 100 + [1] * 100 + [2] * 300
-
-
-@pytest.mark.parametrize(
-    "y",
-    [
-        pytest.param(target, id="list"),
-        pytest.param(pandas.Series(target), id="pandas-series"),
-        pytest.param(numpy.array(target), id="numpy"),
-        pytest.param(numpy.array(target).reshape(-1, 1), id="numpy-column"),
-    ],
-)
-def test_check_high_class_imbalance(y):
-    warning = HighClassImbalanceWarning.check(
-        y=y,
-        stratify=None,
-        ml_task="multiclass-classification",
-    )
-
-    assert warning == HighClassImbalanceWarning.MSG
